refactor: 패키징이동 및 개행적용
