+++ src/main/java/chess/domain/board/Rank.java
@@ -1,8 +1,8 @@
-package domain;
+package chess.domain.board;
 
 import java.util.Arrays;
 
-import static domain.Error.WRONG_RANK_NAME;
+import static chess.domain.additional.Error.WRONG_RANK_NAME;
 
 public enum Rank {
 
@@ -21,14 +21,14 @@ public enum Rank {
         this.name = name;
     }
 
+    public boolean isSameRank(final char name) {
+        return this.name == name;
+    }
+
     public static Rank getRank(final char name) {
         return Arrays.stream(Rank.values())
             .filter(rank -> rank.name == (name))
             .findAny().orElseThrow(() -> new IllegalStateException(WRONG_RANK_NAME.getMessage()));
-    }
-
-    public boolean isSameRank(final char name) {
-        return this.name == name;
     }
 
     public char getName() {
